+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "2×2×2 rubik’s cube movements".
 - Tests for "The polish dictionary".
 - Tests for "1010(1)".
+- Tests for "Vote counting".
 
 ## [1.16.0] - 2022-12-31
 ### Added